requirements.txt
asttokens              2.4.1
certifi                2024.8.30
charset-normalizer     3.4.0
colorama               0.4.6
comm                   0.2.2
contourpy              1.3.0
cycler                 0.12.1
debugpy                1.8.8
decorator              5.1.1
exceptiongroup         1.2.2
executing              2.1.0
filelock               3.16.1
filetype               1.2.0
fonttools              4.54.1
fsspec                 2024.10.0
hub-sdk                0.0.15
idna                   3.7
ipykernel              6.29.5
ipython                8.29.0
jedi                   0.19.2
Jinja2                 3.1.4
jupyter_client         8.6.3
jupyter_core           5.7.2
kiwisolver             1.4.7
MarkupSafe             3.0.2
matplotlib             3.9.2
matplotlib-inline      0.1.7
mpmath                 1.3.0
nest-asyncio           1.6.0
networkx               3.4.2
numpy                  2.1.3
opencv-python          4.10.0.84
opencv-python-headless 4.10.0.84
packaging              24.2
pandas                 2.2.3
parso                  0.8.4
pillow                 11.0.0
pip                    21.2.3
platformdirs           4.3.6
prompt_toolkit         3.0.48
psutil                 6.1.0
pure_eval              0.2.3
py-cpuinfo             9.0.0
Pygments               2.18.0
pyparsing              3.2.0
python-dateutil        2.9.0.post0
python-dotenv          1.0.1
pytz                   2024.2
pywin32                308
PyYAML                 6.0.2
pyzmq                  26.2.0
requests               2.32.3
requests-toolbelt      1.0.0
roboflow               1.1.49
scipy                  1.14.1
seaborn                0.13.2
setuptools             57.4.0
six                    1.16.0
stack-data             0.6.3
sympy                  1.13.1
torch                  2.5.1+cu124
torchaudio             2.5.1+cu124
torchvision            0.20.1+cu124
tornado                6.4.1
tqdm                   4.67.0
traitlets              5.14.3
typing_extensions      4.12.2
tzdata                 2024.2
ultralytics            8.3.29
ultralytics-thop       2.0.11
urllib3                2.2.3
wcwidth                0.2.13